feat(runtime): physically move Tool trait to dasclaw_runtime (#672 step 3)
Per ADR-154 §3 + §10 amendment, move the `Tool` trait from
`desktop-client/ironclaw/src/tools/tool.rs` into a new
`crates/dasclaw_runtime/src/tool.rs`. ironclaw side keeps a thin
`pub use dasclaw_runtime::Tool;` re-export, plus the existing
`crate::context::JobContext`-based trait conformance tests.

## Why dasclaw_runtime, not dasclaw_tool

ADR-154 §3.1 originally said the trait should land in `dasclaw_tool`,
based on the assumption that `dasclaw_tool` already depended on
`dasclaw_runtime`. While implementing this slice we found the real
workspace dep direction is the opposite:

- `dasclaw_runtime` depends on `dasclaw_tool` (needs `dasclaw_tool::ToolError`
  for the `from_tool_impl` adapter in `crates/dasclaw_runtime/src/error.rs`).
- `dasclaw_tool` has no dep on `dasclaw_runtime`.

Making `dasclaw_tool` depend on `dasclaw_runtime` would create an
immediate cycle. The clean fix per user direction is to land the trait
in `dasclaw_runtime` next to `JobContextCore` (which `Tool::execute`
already references) and keep `dasclaw_tool` as the lightweight
"implementation-layer types + errors" crate. ADR-154 §10 (new) records
this amendment with full reasoning so future readers see why the trait
sits beside its required vocab.

## Changes

- `crates/dasclaw_runtime/src/tool.rs` (new): verbatim trait body
  including all default methods (`schema`, `risk_level_for`,
  `requires_approval`, `webhook_capability`, etc.). References
  `JobContextCore` via the in-crate path.
- `crates/dasclaw_runtime/src/lib.rs`: `pub mod tool; pub use tool::Tool;`.
- `crates/dasclaw_runtime/Cargo.toml`: add `rust_decimal` (needed by the
  trait's `estimated_cost` default).
- `desktop-client/ironclaw/src/tools/tool.rs`: replace the inline trait
  definition with `pub use dasclaw_runtime::Tool;` plus a module doc
  citing the ADR amendment. The `#[cfg(test)] mod tests { EchoTool }`
  block remains so we keep proving the trait works with ironclaw's
  concrete `JobContext`.
- `docs/plans/architecture-refactor/adr-154-jobcontext-core-trait-split.md`:
  append §10 amendment.

Every existing call site (`crate::tools::tool::Tool`, every `impl Tool
for ...` in `tools/`, the `WasmToolWrapper` impl, `webhooks/mod.rs`
test stubs, every downstream registration in `dasclaw_mcp` and
`dasclaw_lsp`) resolves through the re-export with zero source changes.

## Sources read
- AGENTS.md (Skills 强制使用规范, PR/commit discipline)
- .github/copilot-instructions.md (§3 pipeline, §4 PR checklist, §6 reuse-before-create)
- docs/plans/architecture-refactor/adr-154-jobcontext-core-trait-split.md (§1, §3.1, §3.5, §4, §8, §10 amendment)
- desktop-client/ironclaw/src/tools/tool.rs (trait source)
- crates/dasclaw_runtime/src/lib.rs (target crate layout)
- crates/dasclaw_runtime/src/error.rs (dep direction evidence)
- crates/dasclaw_tool/Cargo.toml + crates/dasclaw_runtime/Cargo.toml (dep direction evidence)

## Verification
```
cargo check -p dasclaw_runtime --tests   # OK
cargo check -p dasclaw --tests           # OK (5m20s)
cargo fmt --all -- --check               # clean
python3.12 scripts/check_no_panics.py --base origin/xClaw   # OK
cargo clippy --no-deps -p dasclaw_runtime --all-targets -- -D warnings   # clean
cargo clippy --no-deps -p dasclaw --all-targets -- -D warnings           # clean (5m57s)
```

Closes #686
Refs #672
Refs #684
