+++ crates/dasclaw_runtime/Cargo.toml
@@ -11,6 +11,7 @@ dasclaw_tool = { path = "../dasclaw_tool" }
 # `secrets::agent_provider::AgentSecrets` adapter (F3.2 phase 2 PR 4c, #641).
 dasclaw_core = { path = "../dasclaw_core" }
 thiserror = "1"
+rust_decimal = { version = "1", features = ["serde", "serde-with-str", "maths"] }
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 chrono = { version = "0.4", features = ["serde"] }
